curriculum-authoring-service/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Database & Cloud
google-cloud-bigquery==3.17.1
google-cloud-firestore==2.14.0
google-auth==2.27.0

# Authentication
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Integration
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
black==24.1.1
isort==5.13.2
